# [performance] Trim stop_times before stop time interpolation