Stop time interpolation works on the entire `stop_times` df, majority of which will be tossed later based on given requested time range.

Trimming down stop_times before passing it to the stop time interpolation step increases performance by a huge amount. 

Here are some evidence of the performance gain on the `pt.paths.generate_summary_graph_elements` step:

**gtfs id** | **current version** | **with trimming**
--|--|--
f-9q9-bart | 9.0s,  97 edges,  50 stops | 3.8s,  97 edges,  50 stops
f-9q9-actransit | 132.0s,  5670 edges,  5050 stops | 59.8s,  5670 edges,  5050 stops
f-9q8y-sfmta | 131.3s,  3806 edges,  3409 stops | 66.7s,  3806 edges,  3409 stops
f-9qb-goldengatetransit | 10.6s,  519 edges,  473 stops | 7.5s,  519 edges,  473 stops
f-9qc-fairfield~ca~us | 5.0s,  280 edges,  245 stops | 3.1s,  280 edges, 245 stops
f-9qc0-soltrans~ca~us | 6.6s,  475 edges, 408 stops | 4.3s,  475 edges,  408 stops
f-9qc-westcat~ca~us | 6.2s,  263 edges, 220 stops | 3.6s,  263 edges,  220 stops
f-9-amtrak*** | 14.5s,  306 edges,  260 stops | 12.0s,  306 edges,  260 stops