@yiyange 2018-07-14
I ran similar process using UrbanAccess. Here is part of the comparison if you are interested

**gtfs id** | **UrbanAccess** | **PT with trimming**
--|--|--
f-9q9-bart | 4.0s, 97 edges, 50 stops | 3.8s,  97 edges,  50 stops
f-9q9-actransit | Failed | 59.8s,  5670 edges,  5050 stops
f-9q8y-sfmta | 97.4s, 3806 edges, 3409 stops | 66.7s,  3806 edges,  3409 stops
f-9qb-goldengatetransit | Failed | 7.5s,  519 edges,  473 stops
f-9qc-fairfield~ca~us | 4.1s, 297 unique edges, 250 stops | 3.1s,  280 edges, 245 stops
f-9qc0-soltrans~ca~us | 3.2s, 473 unique edges, 407 stops | 4.3s,  475 edges,  408 stops
f-9qc-westcat~ca~us | 2.6s, 263 edges, 220 stops | 3.6s,  263 edges,  220 stops
f-9-amtrak*** | Failed | 12.0s,  306 edges,  260 stops